音のスクリプトの設置場所を変更した
