+++ Assets/Scripts/Audio/AudioResources.cs
@@ -1,0 +1,23 @@
+﻿/// <summary>
+/// BGMの種類
+/// </summary>
+public enum BGM
+{
+    BGM1 = 0,
+    BGM2 = 1,
+}
+
+/// <summary>
+/// SoundEffectの種類
+/// </summary>
+public enum SoundEffect
+{
+    WidgetSelect = 1001,
+
+    Hit1 = 2001,
+    Hit2 = 2002,
+
+    Shot = 3001,
+
+    Clear = 4001,
+}
