--- Assets/Scripts/Manager/Audio/AudioResources.cs
@@ -1,24 +1,0 @@
-﻿
-    /// <summary>
-    /// BGM
-    /// </summary>
-    public enum BGM
-    {
-        BGM1 = 0,
-        BGM2 = 1,
-    }
-
-    /// <summary>
-    /// SoundEffect
-    /// </summary>
-    public enum SoundEffect
-    {
-        WidgetSelect = 1001,
-       
-        Hit1 = 2001,
-        Hit2 = 2002,
-        
-        Shot = 3001,
-        
-        Clear = 4001,
-    }
